# app/back/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.01"
  title: App
  description: API Documentation
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
#host: Taked automatically from site
basePath: /
schemes:
  # tip: remove http to make production-grade
  - http
  #- https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /api/public/auth/login/face:
    x-swagger-router-controller: user.controller
    post:
        description: Authenticate the user with face recognition
        operationId: facelogin
        tags: [Auth, Face, Recognition]
        consumes:
          - application/json
        produces:
          - application/json
        parameters: 
          - name: login
            in: body
            description: Fields to login the user
            required: true
            schema:
              $ref: '#/definitions/faceloginRequest'
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/tokenResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/basicResponse"
...
  /api/public/auth/facebook:
    x-swagger-router-controller: user.controller
    post:
        description: Register/Login the user
        operationId: facebook
        tags: [Auth]
        consumes:
          - application/json
        produces:
          - application/json
        parameters: 
          - name: facebook
            in: body
            description: Fields to create/login the user with Facebook
            required: true
            schema:
              $ref: '#/definitions/facebookRequest'
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/tokenResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/basicResponse"
...
  /api/private/users:
    x-swagger-router-controller: user.controller
    get:
      security:
        - Bearer: []
      description: Get all users
      operationId: getAll
      tags: [Users]
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/usersResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/errorResponse"
    put:
      security:
        - Bearer: []
      description: Update user roles
      operationId: putUserRoles
      tags: [Users]
      consumes:
        - application/json
      produces:
        - application/json
      parameters: 
      - name: user
        in: body
        description: Fields to update the user roles
        required: true
        schema:
          $ref: '#/definitions/userRequest'
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/basicResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/errorResponse"
...
definitions:
...
  tokenResponse:
    type: object
    properties:
      token:
        type: string
...
  faceloginRequest:
    type: object
    properties:
      password:
        type: string
        example: User password
      faceId:
        type: string
        example: Azure Face Id
  userRequest:
    type: object
    properties:
      _id:
        type: string
        example: User id
      email:
        type: string
        example: User email
      displayName:
        type: string
        example: User fullname
      password:
        type: string
        example: User password
      picture:
        type: string
        example: User picture
      provider:
        type: string
        example: User provider
      provider_id:
        type: string
        example: User provider id
      face:
        type: string
        example: Azure Face Id
      roles: 
        type: array
        example: User roles
        items: 
          type: string
  facebookRequest:
    type: object
    properties:
      code:
        type: string
        example: Code
      clientId:
        type: string
        example: ClienId
      GOOGLE_SECRET:
        type: string
        example: Secret
      redirectUri:
        type: string
        example: RedirectUri
...
  errorResponse:
    type: object
    properties:
      message:
        type: string
  basicResponse:
    type: object
    properties:
      message:
        type: string
...
  usersResponse:
    required:
      - list
    properties:
      list:
        type: array
        items:
          $ref: '#/definitions/userResponse'
  userResponse:
    type: object
    required:
      - _id
      - email
      - displayName
      - roles
    properties:
      _id:
        type: string
        example: User id
      email:
        type: string
        example: User email
      displayName:
        type: string
        example: User fullname
      roles: 
        type: array
        example: User roles
        items: 
          type: string